+++ internal/integration/reql_tests/gorethink_test.go
@@ -5,14 +5,19 @@ package reql_tests
 import (
 	"flag"
 	"os"
+	"runtime"
 
-	r "gopkg.in/rethinkdb/rethinkdb-go.v5"
+	r "gopkg.in/rethinkdb/rethinkdb-go.v6"
 )
 
 var url string
 
 func init() {
-	flag.Parse()
+	// Fixing test.testlogfile parsing error on Go 1.13+.
+	if runtime.Version() < "go1.13" {
+		flag.Parse()
+	}
+
 	r.SetVerbose(true)
 
 	// If the test is being run by wercker look for the rethink url