+++ internal/integration/reql_tests/reql_timeout_test.go
@@ -9,8 +9,8 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/suite"
-	r "gopkg.in/rethinkdb/rethinkdb-go.v5"
-	"gopkg.in/rethinkdb/rethinkdb-go.v5/internal/compare"
+	r "gopkg.in/rethinkdb/rethinkdb-go.v6"
+	"gopkg.in/rethinkdb/rethinkdb-go.v6/internal/compare"
 )
 
 // Tests timeouts.
@@ -59,6 +59,7 @@ func (suite *TimeoutSuite) TestCases() {
 	suite.T().Log("Running TimeoutSuite: Tests timeouts.")
 
 	{
+		// r.JS default timeout is 5 sec
 		// timeout.yaml line #5
 		/* err("ReqlQueryLogicError", "JavaScript query `while(true) {}` timed out after 5.000 seconds.", [0]) */
 		var expected_ Err = err("ReqlQueryLogicError", "JavaScript query `while(true) {}` timed out after 5.000 seconds.")