+++ internal/integration/tests/example_connect_test.go
@@ -4,7 +4,7 @@ import (
 	"log"
 	"os"
 
-	r "gopkg.in/rethinkdb/rethinkdb-go.v5"
+	r "gopkg.in/rethinkdb/rethinkdb-go.v6"
 )
 
 var sessionEx *r.Session