+++ internal/integration/tests/query_test.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	test "gopkg.in/check.v1"
-	r "gopkg.in/rethinkdb/rethinkdb-go.v5"
+	r "gopkg.in/rethinkdb/rethinkdb-go.v6"
 )
 
 func (s *RethinkSuite) TestQueryRun(c *test.C) {