+++ internal/integration/tests/rethinkdb_test.go
@@ -6,11 +6,12 @@ import (
 	"fmt"
 	"math/rand"
 	"os"
+	"runtime"
 	"testing"
 	"time"
 
 	test "gopkg.in/check.v1"
-	r "gopkg.in/rethinkdb/rethinkdb-go.v5"
+	r "gopkg.in/rethinkdb/rethinkdb-go.v6"
 )
 
 var session *r.Session
@@ -18,7 +19,11 @@ var testdata = flag.Bool("rethinkdb.testdata", true, "create test data")
 var url, url1, url2, url3, db, authKey string
 
 func init() {
-	flag.Parse()
+	// Fixing test.testlogfile parsing error on Go 1.13+.
+	if runtime.Version() < "go1.13" {
+		flag.Parse()
+	}
+
 	r.SetVerbose(true)
 
 	// If the test is being run by wercker look for the rethink url